Add Critical failure if no default lpad in config
If the user tries to duplicate a computer to use with the `fireworks`
scheduler when no default launchpad has been configured, this currently
fails with an error message that is difficult to understand. Here we simply
add a check that there is a default launchpad configured, and echo a
critical failure that is more helpful in case there hasn't.